+++ ext/dom/attr.c
@@ -60,7 +60,7 @@ PHP_METHOD(domattr, __construct)
 	xmlNodePtr oldnode = NULL;
 	dom_object *intern;
 	char *name, *value = NULL;
-	int name_len, value_len, name_valid;
+	size_t name_len, value_len, name_valid;
 	zend_error_handling error_handling;
 
 	zend_replace_error_handling(EH_THROW, dom_domexception_class_entry, &error_handling TSRMLS_CC);
