+++ ext/dom/entityreference.c
@@ -53,7 +53,7 @@ PHP_METHOD(domentityreference, __construct)
 	xmlNodePtr oldnode = NULL;
 	dom_object *intern;
 	char *name;
-	int name_len, name_valid;
+	size_t name_len, name_valid;
 	zend_error_handling error_handling;
 
 	zend_replace_error_handling(EH_THROW, dom_domexception_class_entry, &error_handling TSRMLS_CC);
