+++ ext/intl/calendar/calendar_methods.cpp
@@ -51,7 +51,7 @@ U_CFUNC PHP_FUNCTION(intlcal_create_instance)
 {
 	zval		*zv_timezone	= NULL;
 	const char	*locale_str		= NULL;
-	int			dummy;
+	size_t			dummy;
 	TimeZone	*timeZone;
 	UErrorCode	status			= U_ZERO_ERROR;
 	intl_error_reset(NULL TSRMLS_CC);
@@ -143,7 +143,7 @@ U_CFUNC PHP_FUNCTION(intlcal_get_keyword_values_for_locale)
 	UErrorCode	status = U_ZERO_ERROR;
 	char		*key,
 				*locale;
-	int			key_len,
+	size_t			key_len,
 				locale_len;
 	zend_bool	commonly_used;
 	intl_error_reset(NULL TSRMLS_CC);
@@ -1144,7 +1144,7 @@ U_CFUNC PHP_FUNCTION(intlcal_from_date_time)
 					zv_timestamp;
 	php_date_obj	*datetime;
 	char			*locale_str			= NULL;
-	int				locale_str_len;
+	size_t				locale_str_len;
 	TimeZone		*timeZone;
 	UErrorCode		status				= U_ZERO_ERROR;
 	Calendar        *cal;
