+++ ext/intl/msgformat/msgformat_format.c
@@ -101,9 +101,9 @@ PHP_FUNCTION( msgfmt_format_message )
 	UChar      *spattern = NULL;
 	int         spattern_len = 0;
 	char       *pattern = NULL;
-	int         pattern_len = 0;
+	size_t         pattern_len = 0;
 	const char *slocale = NULL;
-	int         slocale_len = 0;
+	size_t         slocale_len = 0;
 	MessageFormatter_object mf = {0};
 	MessageFormatter_object *mfo = &mf;
